Refactors ID types to Long and enforces validation
Unifies entity and DTO ID fields to use Long for consistency and future scalability.
Applies validation annotations to request bodies in controllers for improved data integrity.
Standardizes bulk endpoint naming and improves MapStruct configuration to ignore unmapped fields, reducing build warnings.
Adds security user credentials to application properties for easier development setup.

+++ src/main/java/com/portfolio/BiblioHub/category/service/CategoryService.java
@@ -8,13 +8,13 @@ import java.util.List;
 public interface CategoryService {
     CategoryResponseDto createCategory(CategoryRequestDto dto);
 
-    CategoryResponseDto getCategoryById(Integer id);
+    CategoryResponseDto getCategoryById(Long id);
 
     List<CategoryResponseDto> getAllCategories();
 
-    CategoryResponseDto updateCategory(Integer id, CategoryRequestDto dto);
+    CategoryResponseDto updateCategory(Long id, CategoryRequestDto dto);
 
-    void deleteCategory(Integer id);
+    void deleteCategory(Long id);
 
     List<CategoryResponseDto> addAll(List<CategoryRequestDto> dtos);
